Able to generate menu types csv
