+++ server/db/schema.sql
@@ -1,5 +1,4 @@
-CREATE TABLE Restaurants
-(
+CREATE TABLE Restaurants (
   id INT PRIMARY KEY,
   name TEXT,
   top_tags TEXT[],
@@ -9,20 +8,18 @@ CREATE TABLE Restaurants
   join_id INT
 );
 
-  CREATE TABLE Menu
-  (
-    id INT PRIMARY KEY,
-    type TEXT,
-    join_id INT
-  );
+CREATE TABLE Menu (
+  id INT PRIMARY KEY,
+  type TEXT,
+  join_id INT
+);
 
-  CREATE TABLE Dishes
-  (
-    id INT PRIMARY KEY,
-    res_id INT REFERENCES Restaurant(id),
-    menu_id INT REFERENCES Menu(id),
-    dish_name TEXT,
-    dish_info TEXT,
-    price INT,
-    subMenu_type TEXT
-  );
+CREATE TABLE Dishes (
+  id INT PRIMARY KEY,
+  res_id INT REFERENCES Restaurant(id),
+  menu_id INT REFERENCES Menu(id),
+  dish_name TEXT,
+  dish_info TEXT,
+  price INT,
+  subMenu_type TEXT
+);
